STT FKIE/mqtt-transport-master/requirements.txt
flask
paho-mqtt
#Flask-MQTT

apiflask
requests
